Move core modules into autoflow package
