+++ autoflow/gui.py
@@ -6,7 +6,7 @@ _GUI_IMPORTS = {"PyQt5", "pyvistaqt", "matplotlib", "pyvista", "vtk"}
 
 def launch_gui() -> None:
     try:
-        from app import main as app_main
+        from .app import main as app_main
     except ModuleNotFoundError as exc:
         module_name = exc.name or ""
         base_name = module_name.split(".", 1)[0]
