+++ autoflow/io_utils.py
@@ -1,5 +1,5 @@
 import json
-from models import Workspace
+from .models import Workspace
 
 
 def save_workspace_file(path, workspace):
